Bridge protocol_id u16 widening; FCFS registry policy; Rust bridge codec; first-wave protocol map assignments
protocol_id widened u8->u16 (2-octet big-endian) across the wire grammar (CDDL), the protocol registry, and the Go/Rust implementations. Bridge protocol_id registry assignment policy set to First Come First Served (RFC 8126) across the registry CSV, its JSON-schema, docs/registries, GOVERNANCE, and the NEP process. New Rust BridgeEnvelope codec at byte parity with the Go reference on the Bridge channel (decode/encode/correlate). First-wave protocol_id assignments 0x06-0x0A (NLIP, ANP, AGNTCY, AP2, x402) recorded in the registry and the per-protocol mapping documents. Conformance corpus and pins updated to match.

+++ harness/adapters/go/main.go
@@ -419,7 +419,7 @@ func handle(req request) response {
 			return response{Error: "bad foreign hex"}
 		}
 		env := npamp.BridgeEnvelope{
-			Protocol:      npamp.BridgeProtocol(uint8(i(fields, "protocol_id"))),
+			Protocol:      npamp.BridgeProtocol(uint16(i(fields, "protocol_id"))),
 			Kind:          npamp.BridgeMessageKind(uint8(i(fields, "message_kind"))),
 			ContentType:   npamp.BridgeContentType(uint8(i(fields, "content_type"))),
 			Flags:         uint8(i(fields, "flags")),
